generalized concepts are now exported in the appropriate module
i.e. model and parser, respectively
