+++ api/src/parser.rs
@@ -2,6 +2,8 @@
 
 use crate::model::{Quad, Triple};
 use std::error::Error;
+#[cfg(feature = "generalized")]
+pub use crate::generalized::parser::*;
 
 /// A parser returning [`Triple`](../model/struct.Triple.html).
 pub trait TriplesParser: Sized {
